Update readme; minor renaming fix
